Moved Python Type recovery implementation to pysrc2cpg
