+++ console/build.sbt
@@ -13,6 +13,7 @@ dependsOn(
   Projects.semanticcpg,
   Projects.macros,
   Projects.jssrc2cpg,
+  Projects.pysrc2cpg,
   Projects.c2cpg % Test,
   Projects.x2cpg % "compile->compile;test->test"
 )
