+++ joern-cli/frontends/pysrc2cpg/src/main/scala/io/joern/pysrc2cpg/PythonTypeHintCallLinker.scala
@@ -1,4 +1,4 @@
-package io.joern.x2cpg.passes.frontend.impl
+package io.joern.pysrc2cpg
 
 import io.joern.x2cpg.passes.frontend.XTypeHintCallLinker
 import io.shiftleft.codepropertygraph.Cpg
